global: cleanup, remove unused structures and functions; create errno.h and time.h consisting of definitions previously located in posix.h
