--- moose/errno.c
@@ -1,1 +1,0 @@
-int errno;
