+++ moose/fs/ext2.c
@@ -1,9 +1,9 @@
-#include <fs/ext2.h>
-
 #include <assert.h>
 #include <bitops.h>
 #include <blk_device.h>
 #include <endian.h>
+#include <errno.h>
+#include <fs/ext2.h>
 #include <kstdio.h>
 #include <mm/kmalloc.h>
 #include <string.h>
