+++ moose/fs/vfs.c
@@ -1,6 +1,6 @@
 #include "fs/vfs.h"
-
 #include <assert.h>
+#include <errno.h>
 #include <kstdio.h>
 #include <mm/kmalloc.h>
