+++ moose/include/arch/cpu.h
@@ -4,4 +4,3 @@
 #include <types.h>
 
 void dump_registers(void);
-void set_stack(u64 sp, u64 old_stack_base);
