--- moose/include/shell.h
@@ -1,4 +1,0 @@
-#pragma once
-
-int init_shell(void);
-void shell_execute_command(const char *cmd);
