+++ moose/net/frame.c
@@ -1,4 +1,5 @@
 #include <assert.h>
+#include <errno.h>
 #include <fs/posix.h>
 #include <mm/kmalloc.h>
 #include <net/frame.h>
